% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ug.R
\name{theme_ug}
\alias{theme_ug}
\title{Theme Uganda}
\usage{
theme_ug()
}
\value{
Standardised theme for Ugandas plots
}
\description{
More explanations
}
\examples{
iris \%>\% ggplot(aes(x = Sepal.Length, y = Sepal.Width,color = Species)) +
          geom_point() +
          labs(x = "Sepal Length",y = "Sepal Width",
          title = "The length of the Sepals \n  against their width",
          caption = "Source: ggplot2 \n Part of the tidyverse")+
          theme_ug()
}
