% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ug.R
\name{theme_ug}
\alias{theme_ug}
\title{Theme Ugnda}
\usage{
theme_ug()
}
\value{
A standardised Visualalisation
}
\description{
Theme Ugnda
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
geom_point() +
labs(title = "Fuel economy declines as weight increases")
}
